requirements.txt
# Dependencias principales
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-genai
python-dotenv
pdfminer.six
python-docx
requests

# Herramientas de empaquetado
pyinstaller>=6.0.0
